Organiza endpoints do GeraLanding por etapa
