+++ backend/ads-service/src/test/java/com/marketinghub/geralanding/GeraLandingInternalControllerTest.java
@@ -1,5 +1,6 @@
 package com.marketinghub.geralanding;
 
+import com.marketinghub.geralanding.internal.web.GeraLandingInternalController;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
